requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.3
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pytest-django==4.9.0
